Часть проекта, относящаяся к ПЛК, адаптирована для контроллера "BME H58 6040" и новый проект АО "НА". Учтена ситуация с задержкой ответов по Modbus TCP, добавлен функционал, гарантирующий соответствие идентификатора транзакции запроса и ответа. Добавлен функционал, обеспечивающий переключение групп аналоговых параметров в буфере истории (одна группа на один агрегат, один резервуар и т.п.).

+++ iface_rtumaster.c
@@ -23,7 +23,7 @@ void *iface_rtu_master(void *arg)
 	u8  rsp_adu[MB_UNIVERSAL_ADU_LEN];
 	u16 rsp_adu_len;
 
-  int port_id=((unsigned)arg)>>8;
+  int port_id=((long)arg)>>8;
   int client_id, device_id;
 
 	int status;
